Fb master mvp (#55)
* Create dashboard, and change map structure

* Change stucture of main template, so the sidebar is always as long as the whole page

* Add the board positions to the LDAPGroupEnum

* Update TestDataRunner so it will generate more and more relalistic data

* Add chart.js to buidle.gradle

* Reformat code using Intelji reformat

* Improve code for the DashboardController

* Add event sold count when order status is updated

* Reformat of the code

* --wip-- [skip ci]

* Make ready voor AV3

* Add ServiceTest and service tests

* Remove random g

* Fix broken test by making ServiceTest abstract

* Add OrderServiceTest

* Add EventServiceTest

* Add Dockerfile

* Add test properties

* Refactor creation and edit of Events

* Small style changes and setup for sub products

* Fix broken test

* Fb master sub products (#62)

* Start implementing subproducts

* Implement subproducts and fix comments

* Fix broken test

+++ src/main/java/ch/wisv/events/core/model/product/Search.java
@@ -44,7 +44,7 @@ public class Search {
      * @param title Title of the Product
      * @param key   Key of the Product
      */
-    public void addItem(String title, Long key) {
+    public void addItem(String title, Integer key) {
         SearchItem temp = new SearchItem(title, key);
         this.suggestions.add(temp);
     }
